[ 스프린트 10차] (#21)
* refactor: user 엔티티 삭제

* feat: 로그인 인터셉트, 알규먼트 리졸버 구현

* refactor: 클래스 이름 UserAccount -> User 로 수정

* refactor: 클래스 이름 UserAccount -> User 로 수정

* style: 코드 포맷팅 수정

* style: 변수 이름 수정

* refactor: 세션에 id대신 객체 저장

* feat: 노션 폴더 읽기 및 생성 api에 유저 권한 추가

* feat: 픽스쳐 상수 이름 수정 및 노션 폴더 분리 및 병합 api 유저 권한 추가

* test: 노션폴더 소유자가 아닌경우 예외 발생 테스트 코드 추가

* feat: 노션폴더 삭제 및 열람 api에 유저 권한 추가

* refactor: 노션폴더 그래프 조회 및 노션폴더이름 수정 api에 유저 권한 추가

* refactor: 노션 및 노션관계 crud api에 유저 권한 추가

* fix: 로그인 인터셉터 필터링 수정

* fix: 로그인 인터셉터 필터링 수정

* fix: 로그인 인터셉터 필터링 수정

* fix: cors 쿠키 옵션 설정

* fix: yml 설정 수정

* fix: 예외 핸들링

* fix: 세션 쿠키 설정

* fix: 세션 쿠키 설정

* fix: 세션 쿠키 설정

* fix: cors 설정

* fix: cors 설정

* fix: cors 설정

* fix: cors 설정

* fix: cors 설정

* fix: cors 설정

* fix: cors 설정

* fix: cors 설정

* fix: cors 설정

* fix: cookie 설정

* fix: cookie 설정

* fix: cookie 설정

* fix: cookie 설정

* fix: cors 설정

* feat: encrypt password

+++ src/main/java/rmap/service/NotionService.java
@@ -7,6 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import rmap.entity.Edge;
 import rmap.entity.Notion;
 import rmap.entity.NotionFolder;
+import rmap.entity.User;
 import rmap.repository.EdgeRepository;
 import rmap.repository.NotionFolderRepository;
 import rmap.repository.NotionRepository;
@@ -22,29 +23,33 @@ public class NotionService {
     private final EdgeRepository edgeRepository;
     private final NotionFolderRepository notionFolderRepository;
 
-    public NotionResponse readNotion(Long notionId) {
+    public NotionResponse openNotion(User loginedUser, Long notionId) {
         Notion notion = notionRepository.findByIdOrThrow(notionId);
+        Validator.validateNotionOwner(notion, loginedUser);
         return NotionResponse.from(notion);
     }
 
-    public NotionIdResponse createInitialNotion(BuildNotionRequest request) {
+    public NotionIdResponse createInitialNotion(User loginedUser, BuildNotionRequest request) {
         NotionFolder notionFolder = notionFolderRepository.findByIdOrThrow(request.getNotionFolderId());
+        Validator.validateNotionFolderOwner(notionFolder, loginedUser);
         Notion notion = new Notion(request.getName(), request.getContent(), notionFolder);
         return new NotionIdResponse(notionRepository.save(notion).getId());
     }
 
     @Transactional
-    public void demolishNotion(Long notionId) {
+    public void editNotion(User loginedUser, Long notionId, String notionName, String content) {
         Notion notion = notionRepository.findByIdOrThrow(notionId);
-        List<Edge> edges = edgeRepository.findAllByNotionId(notion.getId());
-        edgeRepository.deleteAllInBatch(edges);
-        notionRepository.delete(notion);
+        Validator.validateNotionOwner(notion, loginedUser);
+        notion.editName(notionName);
+        notion.editContent(content);
     }
 
     @Transactional
-    public void editNotion(Long notionId, String name, String content) {
+    public void demolishNotion(User loginedUser, Long notionId) {
         Notion notion = notionRepository.findByIdOrThrow(notionId);
-        notion.editName(name);
-        notion.editContent(content);
+        Validator.validateNotionOwner(notion, loginedUser);
+        List<Edge> edges = edgeRepository.findAllByNotionId(notion.getId());
+        edgeRepository.deleteAllInBatch(edges);
+        notionRepository.delete(notion);
     }
 }
